New folder/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pydantic-settings>=2.6.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.0
scikit-learn>=1.5.0
torch>=2.6.0
torchvision>=0.21.0
pillow>=10.0.0
httpx>=0.27.0
